+++ test/logly_test.dart
@@ -12,7 +12,8 @@ void main() {
   });
   test('level minimums are enforced', () {
     log.level = LogLevel.SHOUT;
-    log.severe("ERROR: LogLevel.SEVERE should not print! ${LogLevel.WARNING >= Logger.root.level}");
+    log.severe("ERROR: LogLevel.SEVERE should not print! "
+        "${LogLevel.WARNING >= Logger.root.level}");
     expect(log.isLoggable(LogLevel.INFO), false);
     expect(log.isLoggable(LogLevel.SHOUT), true);
     log.level = LogLevel.ALL;